requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioredis==2.0.1
aiosignal==1.3.1
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
aspose-words==24.11.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
azure-ai-documentintelligence==1.0.1
azure-core==1.30.2
azure-identity==1.17.1
azure-servicebus==7.12.2
azure-storage-blob==12.24.0
beautifulsoup4==4.13.4
bitarray==3.0.0
black==24.8.0
blinker==1.8.2
Brotli==1.1.0
bytecode==0.16.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
clear==2.0.0
click==8.2.1
cobble==0.1.4
cohere==5.13.12
colbert-ai==0.2.21
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
ddtrace==2.14.2
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.6.1
doclayout_yolo==0.0.3
duckduckgo_search==8.0.2
email_validator==2.2.0
envier==0.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
faiss==1.10.0
fastapi==0.112.1
fastapi-cli==0.0.5
fastapi-restful==0.6.0
fastapi-utils==0.7.0
fastavro==1.10.0
filelock==3.16.1
flake8==7.1.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
fuzzywuzzy==0.18.0
gevent==24.2.1
geventhttpclient==2.3.1
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.0.3
grpcio==1.67.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
identify==2.6.4
idna==3.7
importlib_metadata==8.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-black==0.4.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.45
langchain-experimental==0.3.4
langchain-milvus==0.1.10
langchain-openai==0.2.2
langchain-pymupdf4llm==0.2.0
langchain-text-splitters==0.3.7
langdetect==1.0.9
langgraph==0.3.16
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.17
locust==2.29.1
loguru==0.7.2
lxml==5.4.0
magika==0.6.2
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.1
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
milvus-lite==2.4.12
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.19.2
openai==1.42.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
opentelemetry-api==1.29.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250416
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.2.2
pluggy==1.5.0
plum-dispatch==1.7.4
portalocker==2.10.1
pre_commit==4.0.1
primp==0.15.0
prompt_toolkit==3.0.47
propcache==0.3.0
protobuf==5.29.2
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.1
pyee==11.1.1
pyflakes==3.2.0
Pygments==2.18.0
pygraphviz==1.14
PyJWT==2.10.1
pymilvus==2.5.8
PyMuPDF==1.25.4
pymupdf4llm==0.0.17
pyodbc==5.1.0
pyparsing==3.2.1
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyppeteer==2.0.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==7.4.4
pytest-repeat==0.9.3
pytest-rerunfailures==14.0
python-dateutil==2.9.0
python-dotenv==1.1.0
python-multipart==0.0.9
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
rank-bm25==0.2.2
ray==2.43.0
redis==5.0.8
redisearch==2.0.0
redisvl==0.3.3
referencing==0.36.2
regex==2024.7.24
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rerankers==0.5.3
rich==14.0.0
rmtest==0.7.0
rpds-py==0.23.1
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==75.8.0
shellingham==1.5.4
simsimd==6.2.1
six==1.16.0
smmap==5.0.2
smolagents==1.13.0
sniffio==1.3.1
soupsieve==2.7
spire==0.4.2
Spire.Doc==13.1.0
Spire.Presentation==9.11.2
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.6
stringzilla==3.12.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenize_rt==6.1.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typer==0.15.2
types-requests==2.32.0.20250306
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.29.2
watchfiles==1.0.4
wcwidth==0.2.13
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
XlsxWriter==3.2.3
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
